requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==3.4.7
cytoolz==0.11.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
gunicorn==20.1.0
hexbytes==0.2.1
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.6.3
incremental==21.3.0
ipfshttpclient==0.7.0
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
parsimonious==0.8.1
Pillow==8.3.1
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyrsistent==0.18.0
pytz==2021.1
requests==2.26.0
rlp==2.0.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toolz==0.11.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
varint==1.0.2
web3==5.20.0
websockets==8.1
whitenoise==5.3.0
yarl==1.6.3
zipp==3.4.1
zope.interface==5.4.0
